requirements.txt
# Core dependencies
python>=3.10
numpy>=1.23.0
pandas>=1.5.0

# Geospatial libraries
gdal>=3.0.0
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0
rtree>=1.0.0

# Computer Vision & AI
opencv-python>=4.8.0
pillow>=10.0.0
ultralytics>=8.0.0  # YOLOv11
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0

# Data processing
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional (for development)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0